pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "davybot-plugins-im"
version = "1.0.0"
description = "Official plugin repository for DavyBot Agent Platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "DavyBot Team", email = "to2084@qq.com"}
]
keywords = ["agent", "plugins", "davybot", "gewei"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[tool.setuptools]
packages = []

[tool.pytest.ini_options]
testpaths = ["plugins/*/tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
